did traversing for setting a bit. currently only traverses.
fully testing set relies on bitstringtobzet to work because it is time
consuming to have to manually set up the m_bzet vector before working
on setting/unseeting any bit.
Also: are we going to still store the depth of the tree in the first 8
bits of the bzet vector or are we just keeping it in the m_depth
variable?